add warning about config file naming
put a box around the warning

try to fix the box around the warning

unsure if this is how you make the box work

this table needs less emphasis

that didn't work; will this?

that was all bad, this is good

ok this is definitely it

please

these warnings don't look bad. make the other one look the same